feat: migrate appView.js to ts
